# frontend/docs/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Marketing Campaign Manager API
  description: |
    The Marketing Campaign Manager API provides a comprehensive RESTful interface for managing marketing campaigns. 
    
    ## Features
    - Campaign CRUD operations
    - Metrics tracking and analytics
    - Document and media management
    - Real-time webhooks
    - Comprehensive error handling
    - Rate limiting and authentication
    
    ## Authentication
    This API uses Bearer token authentication. Include your API token in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_TOKEN
    ```
    
    ## Rate Limits
    - Default: 1000 requests per hour
    - Authenticated: 5000 requests per hour
    - Bulk Operations: 100 requests per hour
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@marketing-campaign-manager.com
    url: https://docs.marketing-campaign-manager.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.marketing-campaign-manager.com/v1
    description: Production server
  - url: https://staging-api.marketing-campaign-manager.com/v1
    description: Staging server
  - url: http://localhost:5173/api
    description: Development server
security:
  - bearerAuth: []
paths:
  /campaigns:
    get:
      summary: List campaigns
      description: Retrieve a paginated list of campaigns with optional filtering and sorting
      tags:
        - Campaigns
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Filter by campaign status
          schema:
            type: string
            enum: [Active, Paused, Completed, Draft]
        - name: vendor
          in: query
          description: Filter by vendor name
          schema:
            type: string
        - name: startDate
          in: query
          description: Filter by start date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter by end date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: search
          in: query
          description: Search in campaign names and IDs
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, startDate, revenue, cost, roas]
            default: name
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per hour
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in current window
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp when rate limit resets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create campaign
      description: Create a new marketing campaign
      tags:
        - Campaigns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
            examples:
              basic:
                summary: Basic campaign
                value:
                  name: "Holiday Sale 2024"
                  vendor: "Google Ads"
                  status: "Draft"
                  startDate: "2024-12-01T00:00:00Z"
                  endDate: "2024-12-31T23:59:59Z"
                  manager: "John Doe"
              detailed:
                summary: Detailed campaign
                value:
                  name: "Black Friday 2024"
                  status: "Active"
                  vendor: "Facebook Ads"
                  startDate: "2024-11-25T00:00:00Z"
                  endDate: "2024-11-30T23:59:59Z"
                  manager: "Jane Smith"
                  adPlacementDomain: "shop.example.com"
                  device: ["desktop", "mobile", "tablet"]
                  targeting:
                    interests: ["shopping", "deals", "fashion"]
                    demographics:
                      ageRange: "18-65"
                      gender: "all"
                    location: ["United States", "Canada"]
                  repContactInfo:
                    name: "Mike Johnson"
                    email: "mike@facebook.com"
                    phone: "+1-555-9876"
      responses:
        '201':
          description: Campaign created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /campaigns/{campaignId}:
    get:
      summary: Get campaign by ID
      description: Retrieve detailed information about a specific campaign
      tags:
        - Campaigns
      parameters:
        - name: campaignId
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
            pattern: '^camp_[a-zA-Z0-9_-]+$'
      responses:
        '200':
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update campaign
      description: Update an existing campaign (partial updates supported)
      tags:
        - Campaigns
      parameters:
        - name: campaignId
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
            pattern: '^camp_[a-zA-Z0-9_-]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCampaignRequest'
            examples:
              status_update:
                summary: Update status
                value:
                  status: "Active"
              metrics_update:
                summary: Update metrics
                value:
                  metrics:
                    cost: 1500
                    revenue: 4500
                    rawClicks: 750
                    sales: 45
      responses:
        '200':
          description: Campaign updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignUpdateResponse'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

    delete:
      summary: Delete campaign
      description: Delete a campaign permanently
      tags:
        - Campaigns
      parameters:
        - name: campaignId
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
            pattern: '^camp_[a-zA-Z0-9_-]+$'
      responses:
        '200':
          description: Campaign deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /campaigns/{campaignId}/notes:
    post:
      summary: Add note to campaign
      description: Add a note to a campaign for tracking important information
      tags:
        - Campaign Notes
      parameters:
        - name: campaignId
          in: path
          required: true
          description: Campaign ID
          schema:
            type: string
            pattern: '^camp_[a-zA-Z0-9_-]+$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Note added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CampaignMetrics:
      type: object
      properties:
        cost:
          type: number
          format: float
          minimum: 0
          example: 5000
        revenue:
          type: number
          format: float
          minimum: 0
          example: 15000
        rawClicks:
          type: integer
          minimum: 0
          example: 10000
        uniqueClicks:
          type: integer
          minimum: 0
          example: 8000
        impressions:
          type: integer
          minimum: 0
          example: 100000
        confirmReg:
          type: integer
          minimum: 0
          example: 500
        sales:
          type: integer
          minimum: 0
          example: 200
    Note:
      type: object
      properties:
        id:
          type: string
          pattern: '^note_[a-zA-Z0-9_-]+$'
          example: "note_1234567890"
        text:
          type: string
          maxLength: 1000
          example: "Updated targeting parameters"
        author:
          type: string
          example: "John Doe"
        timestamp:
          type: string
          format: date-time
          example: "2025-01-31T10:45:00Z"
      required:
        - text
        - author
...
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          minimum: 1
          example: 1
        totalPages:
          type: integer
          minimum: 0
          example: 10
        totalItems:
          type: integer
          minimum: 0
          example: 200
        itemsPerPage:
          type: integer
          minimum: 1
          example: 20
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false

    # Request schemas
    CreateCampaignRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        status:
          type: string
          enum: [Active, Paused, Completed, Draft]
          default: Draft
        vendor:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        manager:
          type: string
        adPlacementDomain:
          type: string
        device:
          type: array
          items:
            type: string
            enum: [desktop, mobile, tablet]
        targeting:
          type: object
          properties:
            interests:
              type: array
              items:
                type: string
            demographics:
              type: object
              properties:
                ageRange:
                  type: string
                gender:
                  type: string
                  enum: [all, male, female, other]
            location:
              type: array
              items:
                type: string
        repContactInfo:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
      required:
        - name
        - vendor
        - startDate
        - manager
    UpdateCampaignRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        status:
          type: string
          enum: [Active, Paused, Completed, Draft]
        vendor:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        manager:
          type: string
        adPlacementDomain:
          type: string
        device:
          type: array
          items:
            type: string
            enum: [desktop, mobile, tablet]
        targeting:
          type: object
        repContactInfo:
          type: object
        metrics:
          $ref: '#/components/schemas/CampaignMetrics'
    CreateNoteRequest:
      type: object
      properties:
        text:
          type: string
          maxLength: 1000
        author:
          type: string
      required:
        - text
        - author
...
    CampaignListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            campaigns:
              type: array
              items:
                $ref: '#/components/schemas/Campaign'
            pagination:
              $ref: '#/components/schemas/Pagination'
    CampaignResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            campaign:
              $ref: '#/components/schemas/Campaign'
    CampaignUpdateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            campaign:
              $ref: '#/components/schemas/Campaign'
            changes:
              type: object
              additionalProperties:
                type: object
                properties:
                  old:
                    description: Previous value
                  new:
                    description: New value
    NoteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            note:
              $ref: '#/components/schemas/Note'
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
              example: "Resource deleted successfully"
            deletedId:
              type: string
              example: "camp_1234567890"
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            id:
              type: string
              example: "error_1234567890_abc123"
            category:
              type: string
              enum: [network, authentication, authorization, validation, server, client, timeout, rate_limit, maintenance]
              example: "validation"
            severity:
              type: string
              enum: [low, medium, high, critical]
              example: "low"
            message:
              type: string
              example: "Please check your input and try again."
            technicalMessage:
              type: string
              example: "Invalid campaign name: must be between 3 and 100 characters"
            timestamp:
              type: string
              format: date-time
              example: "2025-01-31T10:30:00Z"
            context:
              type: object
              properties:
                operation:
                  type: string
                  example: "createCampaign"
                requestId:
                  type: string
                  example: "req_1234567890_xyz789"
            isRetryable:
              type: boolean
              example: false
            retryAfter:
              type: integer
              description: Seconds to wait before retrying (for rate limit errors)
              example: 60
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: object
              properties:
                validationErrors:
                  type: object
                  additionalProperties:
                    type: string
                  example:
                    name: "Campaign name is required"
                    vendor: "Vendor is required"